# apps/management-api/openapi.yml
openapi: 3.0.3
info:
  title: Podverse Management API
  description: Comprehensive OpenAPI specification generated from audited route matrix coverage for Podverse management API.
  version: 2.0.0
servers:
  - url: https://management-api.podverse.fm/api/v2
    description: Production Server
  - url: http://localhost:3100/api/v2
    description: Local Development Server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: pv_mgmt_auth
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Error description
              example: Validation failed
            code:
              type: string
              example: validation_error
    AuthenticationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Authentication failure reason
              example: Unauthorized
            code:
              type: string
              example: unauthorized
    AuthorizationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Authorization failure reason
              example: Forbidden
            code:
              type: string
              example: forbidden
    NotFoundErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            message:
              type: string
              description: Resource not found
              example: Not found
            code:
              type: string
              example: not_found
...
    AdminListResponse:
      type: array
      description: List of all admins
      items:
        $ref: "#/components/schemas/Admin"
...
    AdminUpdateResponse:
      allOf:
        - $ref: "#/components/schemas/Admin"
...
security:
  - bearerAuth: []
  - cookieAuth: []
paths:
...
  /admins:
    get:
      operationId: managementGetAdmins
      tags:
        - admins
      summary: GET /admins
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: requireCrud('admins', 'read')."
      security:
        - bearerAuth: []
        - cookieAuth: []
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminListResponse"
              examples:
                adminList:
                  summary: Admin list response
                  value:
                    - id: 1
                      id_text: admin-01
                      email: admin@podverse.fm
                      username: admin
                      role: superuser
                      permissions:
                        feeds_crud: 15
                        feed_takedown_reasons_crud: 15
                        admins_crud: 15
                        stats_crud: 15
                        billing_prices_crud: 15
                        bucket_crud: 15
                        embed_demo_crud: 15
                      created_at: "2024-01-01T00:00:00Z"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
    post:
      operationId: managementPostAdmins
      tags:
        - admins
      summary: POST /admins
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: requireCrud('admins', 'create')."
      security:
        - bearerAuth: []
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateRequest"
            examples:
              createAdmin:
                summary: Create a standard admin with explicit permissions
                value:
                  email: newadmin@podverse.fm
                  username: newadmin
                  password: securePassword123
                  role_id: standard_admin
                  permissions:
                    feeds_crud: 7
                    feed_takedown_reasons_crud: 3
                    admins_crud: 1
                    stats_crud: 7
                    billing_prices_crud: 1
                    bucket_crud: 1
                    embed_demo_crud: 0
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCreateResponse"
              examples:
                adminCreated:
                  summary: Admin created with invite metadata
                  value:
                    id: 12
                    id_text: newadmin-12
                    email: newadmin@podverse.fm
                    username: newadmin
                    role: standard_admin
                    permissions:
                      feeds_crud: 7
                      feed_takedown_reasons_crud: 3
                      admins_crud: 1
                      stats_crud: 7
                      billing_prices_crud: 1
                      bucket_crud: 1
                      embed_demo_crud: 0
                    created_at: "2026-05-13T12:10:00Z"
                    message: Admin created. Invite link generated.
                    set_password_url: https://management.podverse.fm/admins/redeem-invite-link?token=abc123
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
              examples:
                missingPermission:
                  summary: Caller lacks admins:create permission
                  value:
                    message: Forbidden
                    code: forbidden
                    details:
                      - field: permissions
                        message: Missing required admins:create permission
                        code: insufficient_permission
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
  /admins/{id}:
    delete:
      operationId: managementDeleteAdminsId
      tags:
        - admins
      summary: DELETE /admins/{id}
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: requireCrud('admins', 'delete')."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDeleteResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
    get:
      operationId: managementGetAdminsId
      tags:
        - admins
      summary: GET /admins/{id}
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: requireCrud('admins', 'read')."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGetResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
    patch:
      operationId: managementPatchAdminsId
      tags:
        - admins
      summary: PATCH /admins/{id}
      description: "Auto-generated from route matrix. Authenticated endpoint (cookieAuth or bearerAuth). Permission/auth note: requireCrud('admins', 'update')."
      security:
        - bearerAuth: []
        - cookieAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUpdateResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticationErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthorizationErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
...
  /admins/invite-link/redeem:
    post:
      operationId: managementPostAdminsInviteLinkRedeem
      tags:
        - admins
      summary: POST /admins/invite-link/redeem
      description: "Auto-generated from route matrix. Public endpoint. Permission/auth note: public."
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminInviteLinkRedeemRequest"
      responses:
        "200":
          description: Successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminInviteLinkRedeemResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InternalServerErrorResponse"
...